# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "title": "Pool API documentation",
        "contact": {},
        "version": "3.0.0"
    },
    "host": "pool3-api.vivaconagua.org",
    "basePath": "/v1",
    "paths": {
...
        "/finances/taking": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/finances/taking"
                ],
                "summary": "Get a List ofTaking",
                "parameters": [
                    {
                        "type": "string",
                        "name": "artistName",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "name": "crewID",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "eventEndFrom",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "eventEndTo",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "eventName",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "name": "eventState",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "fullCount",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "skip",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "sortDirection",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "sortField",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "name": "statusConfirmed",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "name": "statusNoIncome",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "name": "statusNone",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "name": "statusOpen",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "name": "statusWait",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "name": "typeOfEvent",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/vcago.ResponseListed"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "payload": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Taking"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/finances/taking"
                ],
                "summary": "Update aTaking",
                "parameters": [
                    {
                        "description": "Taking Data",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TakingUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/vcago.ResponseUpdated"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "payload": {
                                            "$ref": "#/definitions/models.Taking"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Validation Error",
                        "schema": {
                            "$ref": "#/definitions/vcago.ValidationErrorResponse"
                        }
                    },
                    "404": {
                        "description": "No Document with given ID",
                        "schema": {
                            "$ref": "#/definitions/vcago.MongoNoDocumentErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "creates an Taking object.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/finances/taking"
                ],
                "summary": "Create aTaking",
                "parameters": [
                    {
                        "description": "Taking Data",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TakingCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Taking succsessfully created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/vcago.ResponseCreated"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "payload": {
                                            "$ref": "#/definitions/models.Taking"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Validation Error",
                        "schema": {
                            "$ref": "#/definitions/vcago.ValidationErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Duplicated Key",
                        "schema": {
                            "$ref": "#/definitions/vcago.MongoDuplicatedErrorResponse"
                        }
                    }
                }
            }
        },
        "/finances/taking/{id}": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/finances/taking"
                ],
                "summary": "Get a Taking by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Taking ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/vcago.ResponseSelected"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "payload": {
                                            "$ref": "#/definitions/models.Taking"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "No Document with given ID",
                        "schema": {
                            "$ref": "#/definitions/vcago.MongoNoDocumentErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/finances/taking"
                ],
                "summary": "Delete a Taking by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Taking ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/vcago.ResponseDeleted"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "payload": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "No Document with given ID",
                        "schema": {
                            "$ref": "#/definitions/vcago.MongoNoDocumentErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Activity": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "model_id": {
                    "type": "string"
                },
                "model_type": {
                    "type": "string"
                },
                "modified": {
                    "$ref": "#/definitions/vmod.Modified"
                },
                "status": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/models.UserDatabase"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "models.Crew": {
            "type": "object",
            "properties": {
                "abbreviation": {
                    "type": "string"
                },
                "additional": {
                    "type": "string"
                },
                "asp_selection": {
                    "type": "string"
                },
                "cities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.City"
                    }
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "mailbox_id": {
                    "type": "string"
                },
                "mattermost_username": {
                    "type": "string"
                },
                "modified": {
                    "$ref": "#/definitions/vmod.Modified"
                },
                "name": {
                    "type": "string"
                },
                "organisation": {
                    "$ref": "#/definitions/models.Organisation"
                },
                "organisation_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "models.DepositUnitTaking": {
            "type": "object",
            "properties": {
                "deposit": {
                    "$ref": "#/definitions/models.DepositDatabase"
                },
                "deposit_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "modified": {
                    "$ref": "#/definitions/vmod.Modified"
                },
                "money": {
                    "$ref": "#/definitions/vmod.Money"
                },
                "status": {
                    "type": "string"
                },
                "taking": {
                    "$ref": "#/definitions/models.Taking"
                },
                "taking_id": {
                    "type": "string"
                }
            }
        },
...
        "models.Source": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "modified": {
                    "$ref": "#/definitions/vmod.Modified"
                },
                "money": {
                    "$ref": "#/definitions/vmod.Money"
                },
                "norms": {
                    "type": "string"
                },
                "payment_type": {
                    "type": "string"
                },
                "taking_id": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "models.Taking": {
            "type": "object",
            "properties": {
                "activities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Activity"
                    }
                },
                "comment": {
                    "type": "string"
                },
                "creator": {
                    "$ref": "#/definitions/models.UserDatabase"
                },
                "crew": {
                    "$ref": "#/definitions/models.Crew"
                },
                "crew_id": {
                    "type": "string"
                },
                "date_of_taking": {
                    "type": "integer"
                },
                "deposit_units": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.DepositUnitTaking"
                    }
                },
                "editor_id": {
                    "type": "string"
                },
                "event": {
                    "$ref": "#/definitions/models.Event"
                },
                "id": {
                    "type": "string"
                },
                "modified": {
                    "$ref": "#/definitions/vmod.Modified"
                },
                "money": {
                    "$ref": "#/definitions/vmod.Money"
                },
                "name": {
                    "type": "string"
                },
                "sources": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Source"
                    }
                },
                "state": {
                    "$ref": "#/definitions/models.TakingState"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "models.TakingCreate": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "crew_id": {
                    "type": "string"
                },
                "date_of_taking": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "new_sources": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SourceCreate"
                    }
                }
            }
        },
        "models.TakingState": {
            "type": "object",
            "properties": {
                "confirmed": {
                    "$ref": "#/definitions/vmod.Money"
                },
                "no_income": {
                    "type": "boolean"
                },
                "open": {
                    "$ref": "#/definitions/vmod.Money"
                },
                "wait": {
                    "$ref": "#/definitions/vmod.Money"
                }
            }
        },
        "models.TakingStateUpdate": {
            "type": "object",
            "properties": {
                "no_income": {
                    "type": "boolean"
                }
            }
        },
        "models.TakingUpdate": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "crew_id": {
                    "type": "string"
                },
                "date_of_taking": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "sources": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SourceUpdate"
                    }
                },
                "state": {
                    "$ref": "#/definitions/models.TakingStateUpdate"
                }
            }
        },
...
        "models.UserDatabase": {
            "type": "object",
            "required": [
                "email",
                "first_name",
                "last_name"
            ],
            "properties": {
                "confirmed": {
                    "type": "boolean"
                },
                "country": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "drops_id": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_login_date": {
                    "type": "integer"
                },
                "last_name": {
                    "type": "string"
                },
                "last_update": {
                    "type": "string"
                },
                "mailbox_id": {
                    "type": "string"
                },
                "modified": {
                    "$ref": "#/definitions/vmod.Modified"
                },
                "organisation_id": {
                    "type": "string"
                },
                "privacy_policy": {
                    "type": "boolean"
                },
                "system_roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/vmod.Role"
                    }
                }
            }
        },
...
        "vcago.MongoDuplicatedErrorResponse": {
            "type": "object",
            "properties": {
                "list_size": {
                    "type": "integer"
                },
                "message": {
                    "type": "string",
                    "example": "duplicate key error: \u003ckey\u003e"
                },
                "model": {
                    "type": "string"
                },
                "payload": {},
                "type": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
        "vcago.MongoNoDocumentErrorResponse": {
            "type": "object",
            "properties": {
                "list_size": {
                    "type": "integer"
                },
                "message": {
                    "type": "string",
                    "example": "document not found"
                },
                "model": {
                    "type": "string"
                },
                "payload": {},
                "type": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
...
        "vcago.ResponseCreated": {
            "type": "object",
            "properties": {
                "list_size": {
                    "type": "integer"
                },
                "message": {
                    "type": "string",
                    "example": "successfully_created"
                },
                "model": {
                    "type": "string"
                },
                "payload": {},
                "type": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "vcago.ResponseDeleted": {
            "type": "object",
            "properties": {
                "list_size": {
                    "type": "integer"
                },
                "message": {
                    "type": "string",
                    "example": "successfully_deleted"
                },
                "model": {
                    "type": "string"
                },
                "payload": {},
                "type": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "vcago.ResponseListed": {
            "type": "object",
            "properties": {
                "list_size": {
                    "type": "integer"
                },
                "message": {
                    "type": "string",
                    "example": "successfully_selected"
                },
                "model": {
                    "type": "string"
                },
                "payload": {},
                "type": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "vcago.ResponseSelected": {
            "type": "object",
            "properties": {
                "list_size": {
                    "type": "integer"
                },
                "message": {
                    "type": "string",
                    "example": "successfully_selected"
                },
                "model": {
                    "type": "string"
                },
                "payload": {},
                "type": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "vcago.ResponseUpdated": {
            "type": "object",
            "properties": {
                "list_size": {
                    "type": "integer"
                },
                "message": {
                    "type": "string",
                    "example": "successfully_updated"
                },
                "model": {
                    "type": "string"
                },
                "payload": {},
                "type": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "vcago.ValidationErrorResponse": {
            "type": "object",
            "properties": {
                "list_size": {
                    "type": "integer"
                },
                "message": {
                    "type": "string",
                    "example": "validate error"
                },
                "model": {
                    "type": "string"
                },
                "payload": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "currency is required"
                    ]
                },
                "type": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
...
        "vmod.Modified": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "integer"
                },
                "updated": {
                    "type": "integer"
                }
            }
        },
        "vmod.Money": {
            "description": "Beispiel eines benutzerdefinierten Geldtyps",
            "type": "object",
            "required": [
                "currency"
            ],
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "currency": {
                    "type": "string"
                }
            }
        },
...
        "vmod.Role": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "root": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Bearer key",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "CookieAuth": {
            "type": "apiKey",
            "name": "access_cookie",
            "in": "cookie"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://hydra.vivaconagua.org/oauth/authorize",
            "tokenUrl": "https://hydra.vivaconagua.org/oauth/token"
        }
    }
}